fix migrations

+++ database/migrations/2026_02_20_032343_create_asistencias_table.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('asistencias')) {
+            return;
+        }
         Schema::create('asistencias', function (Blueprint $table) {
             $table->id();
             $table->foreignId('alumno_id')->constrained('alumnos')->onDelete('cascade');
